moltybox-api/requirements.txt
# Web 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0

# 认证安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 工具
httpx==0.26.0
email-validator==2.1.0

# 数据库
psycopg2-binary==2.9.9

# 开发工具
pytest==7.4.4
pytest-asyncio==0.23.3
